+++ packages/nesd/test/ui/emulator/load_rom_toast_test.dart
@@ -44,6 +44,7 @@ class _Harness {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(true);