+++ packages/nesd/test/ui/emulator/nes_controller_run_state_test.dart
@@ -54,6 +54,7 @@ void main() {
       when(() => settings.region).thenReturn(null);
       when(() => settings.rewind).thenReturn(false);
       when(() => settings.volume).thenReturn(1.0);
+      when(() => settings.lowPassFilter).thenReturn(false);
       when(() => settings.autoSave).thenReturn(false);
       when(() => settings.autoSaveInterval).thenReturn(5);
       when(() => settings.autoLoad).thenReturn(false);
@@ -143,6 +144,7 @@ void main() {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);