+++ packages/nesd/test/ui/emulator/select_rom_route_test.dart
@@ -247,6 +247,7 @@ void main() {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);
@@ -343,6 +344,7 @@ void main() {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);
@@ -396,6 +398,7 @@ void main() {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);
@@ -470,6 +473,7 @@ void main() {
     when(() => settings.region).thenReturn(null);
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
+    when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);